+++ SLHCUpgradeSimulations/L1TrackTrigger/plugins/L1Tracking.icc
@@ -97,7 +97,7 @@
       
     for (int j=0;j<ev.nstubs();j++){
 
-      Stub stub=ev.stub(j);
+      L1TStub stub=ev.stub(j);
 
       int simtrackid=ev.simtrackid(stub);
       //int simtrackid=-1;
@@ -150,7 +150,7 @@
 	if (hypot(stub.x(),stub.y())>60) sigmaz=1.44;  //HACK!!!
       }
 
-      L1TStub aStub(simtrackid,stub.iphi(),stub.iz(),layer,ladder,module,stub.x(),stub.y(),stub.z(),sigmax,sigmaz,stub.pt()*stub.ptsign());
+      L1TStub aStub(simtrackid,stub.diphi(),stub.diz(),layer,ladder,module,stub.x(),stub.y(),stub.z(),sigmax,sigmaz,stub.pt()*stub.ptsign());
 
       //cout << "layer ladder module:"<<layer<<" "<<ladder<<" "<<module
       //	   << " "<<aStub.r()<<" "<<aStub.phi()<<" "<<aStub.z()<<endl;
