--- SimDataFormats/SLHC/interface/L1TConstants.hh
@@ -1,41 +1,0 @@
-#ifndef L1TCONSTANTS_H
-#define L1TCONSTANTS_H
-
-//static double DZ=0.0125;
-//static double DX=0.00125;
-
-//static int MAXZ=32768*2;
-//static int MAXX=8192;
-
-//static int BASE=32768*4;
-//static int BASE2=BASE*64;
-//static int BASE3=1024*8;
-//static int BASE4=8192;
-//static int BASE5=32768;
-//static int BASE6=32768;
-//static int BASE7=32768*4;
-//static int BASE8=1024*16;
-
-//static int BASE9=32768*4;
-
-//static int BASE10=32768*16;
-//static int BASE11=8192*4;
-//static int BASE12=65536;
-//static int BASE13=4096;
-
-//static int BASE14=4;
-//static int BASE15=8;
-//static int BASE16=16;
-//static int BASE17=1024;
-
-//static int BASE18=4096;
-//static int BASE19=1024;
-
-//static int BASE20=512;
-
-//static long long int MAXINT=65535*32768;
-
-#endif
-
-
-
